Fixed blank space input bug in field in menu

+++ Employee Database/include/model/Employee.h
@@ -146,7 +146,7 @@ namespace EmployeeDB::Model {
             return performanceMetric;
         }
 
-        void setPerformanceMetric(float metric) {
+        void setPerformanceMetric(double metric) {
             performanceMetric = metric;
         }
